src/requirements/base.txt
Django==4.2.9
django-environ==0.10.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.4.0
drf-yasg==1.21.5
requests==2.31.0
gunicorn==22.0.0
pika==1.3.2
flower==1.2.0
redis==5.0.7
boto3==1.34.141
django-autoslug==1.9.9
psycopg2-binary==2.9.5
celery-progress==0.4

celery==5.4.0
gevent==24.2.1
pybloom-live===4.0.0

django-elasticsearch-dsl-drf==0.22.5
django-elasticsearch-dsl==7.3.0